feat: move build to packages
