+++ packages/nuxt-legacy/rollup.config.js
@@ -1,5 +1,5 @@
 import babel from 'rollup-plugin-babel'
-import config from '../../build/rollup.config'
+import config from '../nuxt-build/rollup.config'
 
 export default config({
   rootDir: __dirname,
