+++ packages/nuxt-start/rollup.config.js
@@ -1,4 +1,4 @@
-import config from '../../build/rollup.config'
+import config from '../nuxt-build/rollup.config'
 
 export default config({
   rootDir: __dirname
